requirements.txt
# Core utilities
requests==2.31.0
python-dotenv==1.0.1
pydantic==1.10.12
numpy==1.26.4
pandas==2.2.2

# Apache Kafka
kafka-python==2.0.2

# Spark (ARM/M1 safe)
pyspark

# Cassandra
cassandra-driver==3.28.0


# ML / NLP
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99

# Visualization / Analysis
matplotlib==3.9.0
seaborn==0.13.2

# Jupyter
jupyter==1.1.1

# Timezones
pendulum==3.0.0

# Play Store Scraper
google-play-scraper==1.2.2

# Airflow (must stay pinned!)
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.7.1
apache-airflow-providers-ftp==3.5.1
apache-airflow-providers-http==4.5.1
apache-airflow-providers-imap==3.3.1
apache-airflow-providers-sqlite==3.4.3

#postgres
psycopg2-binary